+++ src/Model/AbstractModel.php
@@ -58,6 +58,7 @@ abstract class AbstractModel
 	protected $tripleCount=0; //lazy created
 	protected $uniqueIdGenerator=null; // dependency injections
 	protected $droppedFields = array();
+	protected $globalStorage;
 	
 
 	protected static function mergeOptions( array $options1, array $options2 )
@@ -88,7 +89,7 @@ abstract class AbstractModel
 	/**
 	 * Do not call directlty constructor, use fromArray or other factory methodsinstead
 	 */
-    protected function __construct(array $data = array(), array $customOptions = array()) 
+	protected function __construct(array $data =  [], array $customOptions = [], &$globalStorage = null) 
     { 		
  		$options = self::mergeOptions(self::constructOptions(),$customOptions);
 		
@@ -115,6 +116,7 @@ abstract class AbstractModel
 		$this->options = $options;
 		$this->data = $sanitizedData;
 		$this->setIdGenerator(function($data){return uniqid();});
+		$this->globalStorage = &$globalStorage;
     }
 	
 	
@@ -122,18 +124,18 @@ abstract class AbstractModel
 	/**
 	 * Create an instance from an associative array
 	 */
-	public static function fromArray(array $data, array $customOptions = array())
+    public static function fromArray(array $data, array $customOptions = [] , &$globalStorage = null)
 	{
-		return new static($data,$customOptions);
+	    return new static($data,$customOptions,$globalStorage);
 	}
 	
 	
 	/**
 	 * Create an instance from an generic standard object
 	 */
-	public static function fromStdObject( \stdClass $obj, array $customOptions = array())
+	public static function fromStdObject( \stdClass $obj, array $customOptions = array(), &$globalStorage = null)
 	{
-		return static::fromArray((array)$obj);
+	    return static::fromArray((array)$obj,$customOptions,$globalStorage);
 	}